Refactored MouseAndKeyHandler interface to MouseAndKeyHandlerBase class, refactored (JOGL, GLFW, JavaFX)MouseAndKeyHandler to reflect those changes
