+++ src/main/kotlin/graphics/scenery/controls/GLFWMouseAndKeyHandler.kt
@@ -4,35 +4,26 @@ import com.jogamp.newt.event.InputEvent
 import com.jogamp.newt.event.KeyEvent
 import com.jogamp.newt.event.MouseEvent
 import com.jogamp.newt.event.WindowEvent
-import gnu.trove.map.hash.TIntLongHashMap
-import gnu.trove.set.hash.TIntHashSet
-import net.java.games.input.*
+import graphics.scenery.Hub
+import graphics.scenery.utils.ExtractsNatives
 import org.lwjgl.glfw.GLFW.*
 import org.lwjgl.glfw.GLFWCursorPosCallback
 import org.lwjgl.glfw.GLFWKeyCallback
 import org.lwjgl.glfw.GLFWMouseButtonCallback
 import org.lwjgl.glfw.GLFWScrollCallback
-import org.scijava.ui.behaviour.*
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
-import graphics.scenery.Hub
-import graphics.scenery.controls.behaviours.GamepadBehaviour
-import graphics.scenery.utils.ExtractsNatives
-import java.awt.Toolkit
-import java.io.File
-import java.io.FileOutputStream
-import java.nio.file.Files
-import java.util.*
-import java.util.concurrent.ConcurrentHashMap
-import java.util.jar.JarFile
-import kotlin.concurrent.thread
+import org.scijava.ui.behaviour.InputTrigger
 
 /**
  * Input handling class for GLFW-based windows.
  *
  * @author Ulrik Günther <hello@ulrik.is>
  */
-open class GLFWMouseAndKeyHandler(protected var hub: Hub?) : MouseAndKeyHandler, AutoCloseable, ExtractsNatives {
+open class GLFWMouseAndKeyHandler(protected var hub: Hub?) : MouseAndKeyHandlerBase(), AutoCloseable, ExtractsNatives {
+    /** store os name */
+    private var os = ""
+
+    /** scroll speed multiplier to combat OS idiosyncrasies */
+    private var scrollSpeedMultiplier = 1.0f
 
     var cursorCallback = object : GLFWCursorPosCallback() {
         override fun invoke(window: Long, xpos: Double, ypos: Double) {
@@ -130,63 +121,6 @@ open class GLFWMouseAndKeyHandler(protected var hub: Hub?) : MouseAndKeyHandler,
 
     }
 
-    /** slf4j logger for this class */
-    protected var logger: Logger = LoggerFactory.getLogger("InputHandler")
-
-    /** handle to the active controller */
-    private var controller: Controller? = null
-
-    /** handle to the active controller's polling thread */
-    private var controllerThread: Thread? = null
-
-    /** hash map of the controller's components that are currently above [CONTROLLER_DOWN_THRESHOLD] */
-    private var controllerAxisDown: ConcurrentHashMap<Component.Identifier, Float> = ConcurrentHashMap()
-
-    /** hash map for mouse down events */
-    private var mouseButtonsDown: ConcurrentHashMap<Int, Float> = ConcurrentHashMap()
-
-    /** polling interval for the controller */
-    private val CONTROLLER_HEARTBEAT = 5L
-
-    /** threshold over which an axis is considered down */
-    private val CONTROLLER_DOWN_THRESHOLD = 0.95f
-
-    /** the windowing system's set double click interval */
-    private val DOUBLE_CLICK_INTERVAL = getDoubleClickInterval()
-
-    /**
-     * Queries the windowing system for the current double click interval
-     *
-     * @return The double click interval in ms
-     */
-    private fun getDoubleClickInterval(): Int {
-        val prop = Toolkit.getDefaultToolkit().getDesktopProperty("awt.multiClickInterval")
-
-        if (prop == null) {
-            return 200
-        } else {
-            return prop as Int
-        }
-    }
-
-    /** ui-behaviour input trigger map */
-    private var inputMap: InputTriggerMap? = null
-
-    /** ui-behaviour behaviour map */
-    private var behaviourMap: BehaviourMap? = null
-
-    /** expected modifier count */
-    private var inputMapExpectedModCount: Int = 0
-
-    /** behaviour expected modifier count */
-    private var behaviourMapExpectedModCount: Int = 0
-
-    /** store os name */
-    private var os = ""
-
-    /** scroll speed multiplier to combat OS idiosyncrasies */
-    private var scrollSpeedMultiplier = 1.0f
-
     init {
         os = if(System.getProperty("os.name").toLowerCase().indexOf("windows") != -1) {
             "windows"
@@ -204,207 +138,7 @@ open class GLFWMouseAndKeyHandler(protected var hub: Hub?) : MouseAndKeyHandler,
             10.0f
         }
 
-        logger.debug("Native JARs for JInput: ${getNativeJars("jinput-platform", hint = "jinput-raw.dll").joinToString(", ")}")
-        extractLibrariesFromJar(getNativeJars("jinput-platform", hint = "jinput-raw.dll"))
-
-        ControllerEnvironment.getDefaultEnvironment().controllers.forEach {
-            if (it.type == Controller.Type.STICK || it.type == Controller.Type.GAMEPAD) {
-                this.controller = it
-                logger.info("Added gamepad controller: $it")
-            }
-        }
-
-        controllerThread = thread {
-            var event_queue: EventQueue
-            val event: Event = Event()
-
-            while (true) {
-                controller?.let {
-                    controller!!.poll()
-
-                    event_queue = controller!!.eventQueue
-
-                    while (event_queue.getNextEvent(event)) {
-                        controllerEvent(event)
-                    }
-                }
-
-                for (gamepad in gamepads) {
-                    for (it in controllerAxisDown) {
-                        if (Math.abs(it.value) > 0.02f && gamepad.behaviour.axis.contains(it.key)) {
-                            logger.trace("Triggering ${it.key} because axis is down (${it.value})")
-                            gamepad.behaviour.axisEvent(it.key, it.value.toFloat())
-                        }
-                    }
-                }
-
-                Thread.sleep(this.CONTROLLER_HEARTBEAT)
-            }
-        }
     }
-
-    /**
-     * Sets the input trigger map to the given map
-     *
-     * @param[inputMap] The input map to set
-     */
-    override fun setInputMap(inputMap: InputTriggerMap) {
-        this.inputMap = inputMap
-        inputMapExpectedModCount = inputMap.modCount() - 1
-    }
-
-    /**
-     * Sets the behaviour trigger map to the given map
-     *
-     * @param[behaviourMap] The behaviour map to set
-     */
-    override fun setBehaviourMap(behaviourMap: BehaviourMap) {
-        this.behaviourMap = behaviourMap
-        behaviourMapExpectedModCount = behaviourMap.modCount() - 1
-    }
-
-    /**
-     * Managing internal behaviour lists.
-     *
-     * The internal lists only contain entries for Behaviours that can be
-     * actually triggered with the current InputMap, grouped by Behaviour type,
-     * such that hopefully lookup from the event handlers is fast,
-     *
-     * @property[buttons] Buttons triggering the input
-     * @property[behaviour] Behaviour triggered by these buttons
-     */
-    internal class BehaviourEntry<T : Behaviour>(
-        val buttons: InputTrigger,
-        val behaviour: T)
-
-    private val buttonDrags = ArrayList<BehaviourEntry<DragBehaviour>>()
-
-    private val keyDrags = ArrayList<BehaviourEntry<DragBehaviour>>()
-
-    private val buttonClicks = ArrayList<BehaviourEntry<ClickBehaviour>>()
-
-    private val keyClicks = ArrayList<BehaviourEntry<ClickBehaviour>>()
-
-    private val scrolls = ArrayList<BehaviourEntry<ScrollBehaviour>>()
-
-    private val gamepads = ArrayList<BehaviourEntry<GamepadBehaviour>>()
-
-    /**
-     * Make sure that the internal behaviour lists are up to date. For this, we
-     * keep track the modification count of [.inputMap] and
-     * [.behaviourMap]. If expected mod counts are not matched, call
-     * [.updateInternalMaps] to rebuild the internal behaviour lists.
-     */
-    @Synchronized private fun update() {
-        val imc = inputMap!!.modCount()
-        val bmc = behaviourMap!!.modCount()
-
-        if (imc != inputMapExpectedModCount || bmc != behaviourMapExpectedModCount) {
-            inputMapExpectedModCount = imc
-            behaviourMapExpectedModCount = bmc
-            updateInternalMaps()
-        }
-    }
-
-    /**
-     * Build internal lists buttonDrag, keyDrags, etc from BehaviourMap(?) and
-     * InputMap(?). The internal lists only contain entries for Behaviours that
-     * can be actually triggered with the current InputMap, grouped by Behaviour
-     * type, such that hopefully lookup from the event handlers is fast.
-     */
-    private fun updateInternalMaps() {
-        buttonDrags.clear()
-        keyDrags.clear()
-        buttonClicks.clear()
-        keyClicks.clear()
-
-        for (entry in inputMap!!.getAllBindings().entries) {
-            val buttons = entry.key
-            val behaviourKeys = entry.value ?: continue
-
-            for (behaviourKey in behaviourKeys) {
-                val behaviour = behaviourMap!!.get(behaviourKey) ?: continue
-
-                if (behaviour is DragBehaviour) {
-                    val dragEntry = BehaviourEntry<DragBehaviour>(buttons, behaviour)
-                    if (buttons.isKeyTriggered())
-                        keyDrags.add(dragEntry)
-                    else
-                        buttonDrags.add(dragEntry)
-                }
-
-                if (behaviour is ClickBehaviour) {
-                    val clickEntry = BehaviourEntry<ClickBehaviour>(buttons, behaviour)
-                    if (buttons.isKeyTriggered())
-                        keyClicks.add(clickEntry)
-                    else
-                        buttonClicks.add(clickEntry)
-                }
-
-                if (behaviour is ScrollBehaviour) {
-                    val scrollEntry = BehaviourEntry<ScrollBehaviour>(buttons, behaviour)
-                    scrolls.add(scrollEntry)
-                }
-
-                if (behaviour is GamepadBehaviour) {
-                    val gamepadEntry = BehaviourEntry<GamepadBehaviour>(buttons, behaviour)
-                    gamepads.add(gamepadEntry)
-                }
-            }
-        }
-    }
-
-
-    /**
-     * Which keys are currently pressed. This does not include modifier keys
-     * Control, Shift, Alt, AltGr, Meta.
-     */
-    private val pressedKeys = TIntHashSet(5, 0.5f, -1)
-
-    /**
-     * When keys where pressed
-     */
-    private val keyPressTimes = TIntLongHashMap(100, 0.5f, -1, -1)
-
-    /**
-     * Whether the SHIFT key is currently pressed. We need this, because for
-     * mouse-wheel AWT uses the SHIFT_DOWN_MASK to indicate horizontal
-     * scrolling. We keep track of whether the SHIFT key was actually pressed
-     * for disambiguation.
-     */
-    private var shiftPressed = false
-
-    /**
-     * Whether the META key is currently pressed. We need this, because on OS X
-     * AWT sets the META_DOWN_MASK to for right clicks. We keep track of whether
-     * the META key was actually pressed for disambiguation.
-     */
-    private var metaPressed = false
-
-    /**
-     * Whether the WINDOWS key is currently pressed.
-     */
-    private var winPressed = false
-
-    /**
-     * The current mouse coordinates, updated through [.mouseMoved].
-     */
-    private var mouseX: Int = 0
-
-    /**
-     * The current mouse coordinates, updated through [.mouseMoved].
-     */
-    private var mouseY: Int = 0
-
-    /**
-     * Active [DragBehaviour]s initiated by mouse button press.
-     */
-    private val activeButtonDrags = ArrayList<BehaviourEntry<DragBehaviour>>()
-
-    /**
-     * Active [DragBehaviour]s initiated by key press.
-     */
-    private val activeKeyDrags = ArrayList<BehaviourEntry<DragBehaviour>>()
 
     /**
      * Returns the key mask of a given input event
@@ -481,8 +215,8 @@ open class GLFWMouseAndKeyHandler(protected var hub: Hub?) : MouseAndKeyHandler,
     fun mouseMoved(e: MouseEvent) {
         update()
 
-        mouseX = e.getX()
-        mouseY = e.getY()
+        mouseX = e.x
+        mouseY = e.y
 
         for (drag in activeKeyDrags) {
             drag.behaviour.drag(mouseX, mouseY)
@@ -617,11 +351,9 @@ open class GLFWMouseAndKeyHandler(protected var hub: Hub?) : MouseAndKeyHandler,
         }
 
         val clickMask = mask and InputTrigger.DOUBLE_CLICK_MASK.inv()
-        for (click in buttonClicks) {
-            if (click.buttons.matches(mask, pressedKeys) || clickMask != mask && click.buttons.matches(clickMask, pressedKeys)) {
-                click.behaviour.click(x, y)
-            }
-        }
+        buttonClicks
+            .filter { it.buttons.matches(mask, pressedKeys) || clickMask != mask && it.buttons.matches(clickMask, pressedKeys) }
+            .forEach { it.behaviour.click(x, y) }
     }
 
     /**
@@ -725,52 +457,6 @@ open class GLFWMouseAndKeyHandler(protected var hub: Hub?) : MouseAndKeyHandler,
         shiftPressed = false
         metaPressed = false
         winPressed = false
-    }
-
-    /**
-     * Called when a new controller is added
-     *
-     * @param[event] The incoming controller event
-     */
-    override fun controllerAdded(event: ControllerEvent?) {
-        if(controller == null && event != null && event.controller.type == Controller.Type.GAMEPAD) {
-            logger.info("Adding controller ${event.controller}")
-            this.controller = event.controller
-        }
-    }
-
-    /**
-     * Called when a controller is removed
-     *
-     * @param[event] The incoming controller event
-     */
-    override fun controllerRemoved(event: ControllerEvent?) {
-        if(event != null && controller != null) {
-            logger.info("Controller removed: ${event.controller}")
-
-            controller = null
-        }
-    }
-
-    /**
-     * Called when a controller event is fired. This will update the currently down
-     * buttons/axis on the controller.
-     *
-     * @param[event] The incoming controller event
-     */
-    fun controllerEvent(event: Event) {
-        for(gamepad in gamepads) {
-            if(event.component.isAnalog && Math.abs(event.component.pollData) < CONTROLLER_DOWN_THRESHOLD) {
-                logger.trace("${event.component.identifier} over threshold, removing")
-                controllerAxisDown.put(event.component.identifier, 0.0f)
-            } else {
-                controllerAxisDown.put(event.component.identifier, event.component.pollData)
-            }
-
-            if(gamepad.behaviour.axis.contains(event.component.identifier)) {
-                gamepad.behaviour.axisEvent(event.component.identifier, event.component.pollData)
-            }
-        }
     }
 
     override fun close() {
